feat!: upgrade aws provider, upgrade lambda node version
this change came about because the lambda node version 16 is being
removed by aws this year.

version 4 of the aws provider doesn't contain any lambda runtime version
of node above 16, so we have to upgrade the provider.

mark this change as breaking, as it will require users of the provider
to upgrade the aws provider configuration.

additionally, adds a configuration for overriding alert lambda settings.

and removes a hardcoded vpc configuration because it is deprecated and
causes a warning message.