fix!: node group to use only one availability zone
Currently, the node groups are being created with access to all
available subnets, defeating the whole purpose of having different node
groups. Each node group can have its nodes created in any subnet, and
thus any az. I think this negatively impacts the cluster autoscaler's
ability to scale correctly.

I think this was an oversite on my part when I created this part of the
module, and I never realized until now, so I'm correcting it
:sweat_smile:

I'm considering this to be a breaking change as it will cause node
groups to get recreated, which is unexpected behavior.

BREAKING CHANGE: will recreate node groups.